feat(types): add typeRecord/typeThing helpers + canonical surqlFn module (#6)
Introduce src/types/surqlFn.ts as the canonical home for SurrealDB
server-side function values:

- typeRecord(table, id?) / typeThing(table, id?) render `type::record(...)`
  expressions that compose naturally with `.set()`, `.where()`,
  `.select()`, and the CRUD helpers (no more raw-string interpolation).
- surqlFn / isSurqlFn / SurrealFnValue move here; old import sites
  continue to work via back-compat re-exports in query/expressions.ts.
- resolveRecordTarget extracts `table:id` from either a raw string or a
  typeRecord ref, letting downstream helpers accept both shapes.

Public exports live on `@oneiriq/surql` via src/types/mod.ts so callers
can `import { typeRecord, typeThing } from '@oneiriq/surql'`.
