+++ src/query/mod.ts
@@ -56,7 +56,6 @@ export {
   floor,
   func,
   type FunctionExpression,
-  isSurqlFn,
   lower,
   mathMax,
   mathMean,
@@ -70,8 +69,6 @@ export {
   round_,
   stringLength,
   sum_,
-  surqlFn,
-  type SurrealFnValue,
   timeFormat,
   timeNow,
   typeIs,
